% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRadiation.R
\name{calculateRadiation}
\alias{calculateRadiation}
\title{calculateRadiation}
\usage{
calculateRadiation(MNT, zoneEtude, date = as.Date("2019-06-20"))
}
\arguments{
\item{MNT}{Raster du Modele Numerique de Terrain de la zoneEtude obtenu avec la fonction importMNT}

\item{zoneEtude}{Objet sf contenant le shape de la zoneEtude}

\item{date}{Date au format "Years-Month-Day", par exemple "1997-04-24", le plus beau jour de 1997}
}
\value{
Renvoi un raster avec les valeurs de rayonnements
}
\description{
calculateRadiation
}
