% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeCommune.R
\name{codeCommune}
\alias{codeCommune}
\title{codeCommune}
\usage{
codeCommune(path)
}
\arguments{
\item{path}{Adresse où se trouve le fichier d'entrees parcCad.xlsx}
}
\value{
Les codes communes associés aux parcellaires
}
\description{
codeCommune
}
