% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeINSEE.R
\name{codeINSEE}
\alias{codeINSEE}
\title{codeINSEE}
\usage{
codeINSEE(code_postal = "29760", interactive = FALSE)
}
\arguments{
\item{code_postal}{Code postal au format "character"}

\item{interactive}{Si TRUE, permet de selectionner la commune parmis}
}
\value{
Renvoi le code INSEE au format caractere
}
\description{
Permet de convertir un code postal en code INSEE
}
