% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courbeNiveau.R
\name{courbeNiveau}
\alias{courbeNiveau}
\title{courbeNiveau}
\usage{
courbeNiveau(MNT)
}
\arguments{
\item{MNT}{Fichier MNT obtenue avec la fonction importMNT}
}
\value{
Renvoi les courbes de niveau a partir d'un MNT
}
\description{
courbeNiveau
}
