% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_commune.R
\docType{data}
\name{data_commune}
\alias{data_commune}
\title{data_commune}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36742 rows and 5 columns.
}
\usage{
data_commune
}
\description{
data_commune
}
\keyword{datasets}
