% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_departement.R
\docType{data}
\name{data_departement}
\alias{data_departement}
\title{data_departement}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 102 rows and 3 columns.
}
\usage{
data_departement
}
\description{
data_departement
}
\keyword{datasets}
