% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNplacettes.R
\docType{data}
\name{IFNplacettes}
\alias{IFNplacettes}
\title{Donnes concernant les placettesIFN}
\format{
An object of class data.frame
}
\usage{
data(IFNplacettes)
}
\description{
Donnes concernant les placettesIFN
}
\examples{
data(IFNplacettes)

}
\keyword{datasets}
