% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAAC.R
\name{importAAC}
\alias{importAAC}
\title{importAAC}
\usage{
importAAC()
}
\value{
Renvoi une couche sf indiquant l'ensemble des AAC de France
}
\description{
Fonction permettant de telecharger les Aires d'Alimentation de Captage.
Une AAC désigne la zone en surface sur laquelle l’eau qui s’infiltre ou ruisselle alimente le captage.
L’extension de ces surfaces est généralement plus vaste que celle des périmètres de protection de captage.
}
