% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCad.R
\name{importCad}
\alias{importCad}
\title{importCad}
\usage{
importCad(codeINSEE)
}
\arguments{
\item{codeINSEE}{Code INSEE d'une commune de France}
}
\value{
Renvoi les cadastres de la commune selectionnee
}
\description{
importCad
}
