% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGeol.R
\name{importGeol}
\alias{importGeol}
\title{importGeol}
\usage{
importGeol(zoneEtude)
}
\arguments{
\item{zoneEtude}{Objet sf contenant la zone d'etude}
}
\value{
Renvoi la couche geologoque de la zone etudie
}
\description{
importGeol
}
