% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMH.R
\name{importMH}
\alias{importMH}
\title{importMH}
\usage{
importMH()
}
\value{
Renvoi une shape contenant les monuments historique de France
}
\description{
Fonction permettatn de telecharger le shape tout les monuments historique de France
}
