% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMNT.R
\name{importMNT}
\alias{importMNT}
\title{importMNT}
\usage{
importMNT(zoneEtude, buffer = 100, zoom = 10)
}
\arguments{
\item{zoneEtude}{L'objet sf contenant nos parcelles cadastrales}

\item{buffer}{Zone de prospection en metre}

\item{zoom}{Nibveau de precision du MNT. Voir fonction "get_elev_raster" du package elevatr pour plus d'infos.}
}
\value{
Renvoi le MNT de la zoneEtude
}
\description{
importMNT
}
