% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPlacettesIFN.R
\name{importPlacettesIFN}
\alias{importPlacettesIFN}
\title{importPlacettesIFN}
\usage{
importPlacettesIFN(zoneEtude, buffer = 1500)
}
\arguments{
\item{zoneEtude}{L'objet sf contenant nos parcelles cadastrales}

\item{buffer}{Zone de prospection en metre}
}
\value{
Renvoi la position des placettes IFN presentes dans le buffer
}
\description{
importPlacettesIFN
}
