% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTempMean.R
\name{importTempMean}
\alias{importTempMean}
\title{importTempMean}
\usage{
importTempMean(
  zoneEtude,
  firstDate = stop("Choisir une date de depart"),
  endDate = format(Sys.time(), "\%Y-\%m-\%d"),
  bufferTemp = 4000
)
}
\arguments{
\item{zoneEtude}{Objet sf contenant le shape de la zoneEtude}

\item{firstDate}{Premiere prise en compte dans la moyenne}

\item{endDate}{Derniere date prise en compte dans la moyenne. Egale à la date du jour si non indique}

\item{bufferTemp}{Zone de prospection autour de la zone d'etude}
}
\value{
Renvoi un raster contenant les temperatures moyennees sur la periode choisie
}
\description{
Fonction permettant de recuperer les temperatures moyennes a partir des donnees MODIS.
La raster a une resolution de 1km.
}
