% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepOutline.R
\name{keepOutline}
\alias{keepOutline}
\title{Outline}
\usage{
keepOutline(parcelles)
}
\arguments{
\item{parcelles}{L'objet sf contenant les parcelles}
}
\value{
Le contour de nos parcelles. Format sortie : Lambert93 / EPSG : 2154
}
\description{
Outline
}
