% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liens_MNT_IGN.R
\docType{data}
\name{liens_MNT_IGN}
\alias{liens_MNT_IGN}
\title{liens_MNT_IGN}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 208 rows and 3 columns.
}
\usage{
liens_MNT_IGN
}
\description{
liens_MNT_IGN
}
\keyword{datasets}
