% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placettes_IFN.R
\docType{data}
\name{placettes_IFN}
\alias{placettes_IFN}
\title{placettes_IFN}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 87533 rows and 8 columns.
}
\usage{
placettes_IFN
}
\description{
placettes_IFN
}
\keyword{datasets}
