% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAAC.R
\name{plotAAC}
\alias{plotAAC}
\title{PlotAAC}
\usage{
plotAAC(AACShp, zoneEtude, bufferAAC = 1000)
}
\arguments{
\item{AACShp}{Object sf contenant le shp des AAC de France. Ce dernier peut se recuperer avec la fonction importAAC}

\item{zoneEtude}{Objet sf contenant le shape de la zoneEtude}

\item{bufferAAC}{Perimetre de prospection autour de la zoneEtude}
}
\value{
Renvoi une carte interactive avec la zone d'etude et les differentes AAC autour de la zoneEtude
}
\description{
PlotAAC
}
