% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCad.R
\name{plotCad}
\alias{plotCad}
\title{plotCad}
\usage{
plotCad(zoneEtude, printID = TRUE, printPlacIFN = TRUE, bufferIFN = 1500)
}
\arguments{
\item{zoneEtude}{Objet sf contenant nos parcelles}

\item{printID}{Si TRUE, les noms des cadastres sont affichees. Il faut passer la souris sur les points rouges}

\item{printPlacIFN}{Si TRUE, on telecharge les placettes IFN autour de la zone d'etude}

\item{bufferIFN}{Distance de prospection de placettes IFN autour de la zoneEtude}
}
\value{
Une carte interactive permettant de voir l'emplacement de nos cadastres et les placettes IFN prochent
}
\description{
plotCad
}
