% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotETPMean.R
\name{plotETPMean}
\alias{plotETPMean}
\title{plotETPMean}
\usage{
plotETPMean(ETPMeanRaster, zoneEtude)
}
\arguments{
\item{ETPMeanRaster}{Raster des ETP moyennes obtenues avec la fonction importETPMean}

\item{zoneEtude}{Objet sf contenant le shape de la zoneEtude}
}
\value{
Renvoi une carte interactive de la temperature moyennes autour de notre zoneEtude.
}
\description{
plotETPMean
}
