% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMH.R
\name{plotMH}
\alias{plotMH}
\title{plotMH}
\usage{
plotMH(MHShp, zoneEtude, bufferMH = 50000)
}
\arguments{
\item{MHShp}{Objef sf contenant le shape des monuments historiques de France}

\item{zoneEtude}{Objet sf contenant le shape de la zoneEtude}

\item{bufferMH}{Perimetre de prospection autour de la zoneEtude}
}
\value{
Renvoi une carte interactive representant le zoneEtude ainsi que les monuments historiques dans la zoneEtude
}
\description{
plotMH
}
