% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCad.R
\name{plotParcelles}
\alias{plotParcelles}
\title{plotCad}
\usage{
plotParcelles(zoneEtude, printID = TRUE)
}
\arguments{
\item{zoneEtude}{Objet sf contenant nos parcelles}

\item{printID}{Si TRUE, les noms des cadastres sont affichees. Il faut passer la souris sur les points rouges}
}
\value{
Une carte interactive permettant de voir l'emplacement de nos cadastres et les placettes IFN prochent
}
\description{
plotCad
}
