% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectID.R
\name{selectID}
\alias{selectID}
\title{selectID}
\usage{
selectID(zoneEtude, mapBackground = "OpenStreetMap")
}
\arguments{
\item{zoneEtude}{Objet sf contenant les cadastres de la zoneEtude}

\item{mapBackground}{Choix du fond de carte : "OpenStreetMap" , "Scan25" , "Ortho"}
}
\value{
Renvoi un tableau avec les identifiants des zones choisies
}
\description{
selectID
}
