% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectDepartement.R
\name{detectDepartement}
\alias{detectDepartement}
\title{detectDepartement}
\usage{
detectDepartement(shape)
}
\arguments{
\item{shape}{Objet sf contenant un seul polygone}
}
\value{
Renvoi le numero de departement du polygon en entree
}
\description{
Detecte dans quel departement se trouve le polygon en entree
}
