% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCad.R
\name{importCad}
\alias{importCad}
\title{importCad}
\usage{
importCad(codeCommune)
}
\arguments{
\item{codeCommune}{codes communes étudiés. Si il existe un fichier "parcCad.xlsx", les cadastres sont directement triés}
}
\value{
Renvoi les cadastres souhaités
}
\description{
importCad
}
