% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMNT.R
\name{importMNT}
\alias{importMNT}
\title{importMNT}
\usage{
importMNT(zoneEtude, res = "25m", codeDep, source = FALSE)
}
\arguments{
\item{zoneEtude}{objet de class sf corresopndant a un unique polygone}

\item{res}{resolution du raster : "25m" ou "5m"}

\item{codeDep}{si non renseigne, le departement est detecte automatiquement (+ 5/10s). Attention, il faut bien verifier que les polygones en entrees soient TOUS dans le codeDep renseigne.}

\item{source}{si TRUE, la fonction renvoi la source du MNT au format shp}
}
\value{
renvoi le MNT de la zone etudiee ou la source de la donnee si "source" = TRUE
}
\description{
Permet de telecharger les MNT de l'IGN a la resolution 5m et 25m sur toute la France y compris les DOM-TOMs
}
