% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMNT.R
\name{plotMNT}
\alias{plotMNT}
\title{plotMNT}
\usage{
plotMNT(MNT, zoneEtude, distInterCourbe = 100)
}
\arguments{
\item{MNT}{Raster MNT obtenue avec la fonction import MNT}

\item{zoneEtude}{Objet sf comportant le/les zones etudiees}

\item{distInterCourbe}{Altitude entre deux courbes de niveau}
}
\value{
Imprime la carte avec le MNT
}
\description{
plotMNT
}
