% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCad.R
\name{selectID}
\alias{selectID}
\title{selectID}
\usage{
selectID(zoneEtude)
}
\arguments{
\item{zoneEtude}{Objet sf contenant les cadastres de la zoneEtude}
}
\value{
Renvoi un tableau avec les identifiants des zones choisies
}
\description{
selectID
}
